Disable tests that require credits
Tests `testDoStabilityJob`, `testDoMidjourneyJob`, and `getResponseFromJobResponse` in `PicogenTest.java` have been marked as disabled as they require credits to run. These tests can only be executed when credits are available or when a valid id is available for use. Additionally, an instruction was added in `README.md` providing steps to execute `ImageDownloader` class using the `./gradlew run` task.