"Enhance file splitting tests"
This commit deletes a line that was printing source file format information in the WavFileSplitter.

This commit also brings enhancements to the WavFileSplitterTest. Now, the tests are conducted on three file sizes (small, medium, and large) instead of having a single generic large file test. This provides more diverse testing scenarios, aiding in the robustness of the WavFileSplitter. Additional assertions were also added to ensure the total size of split chunks equals the original file size within a 1% deviation.
.