Disable saveImagesToFiles test
The test method saveImagesToFiles in PicogenTest.java has been marked as Disabled. The test was initially used to ensure the proper saving of image files from URLs. It is no longer necessary as the functionality is now working as expected, hence, the test is disabled to improve test suite runtime.