Update file writing method and dependencies
Improved the "writeWordDocument" method in FileUtils for better readability and functionality. Made the method more specific - 'addTitleToWordDocument' and 'addTextToWordDocument' instead of generic 'addTitleToDocument' and 'addTextToDocument'. Only creating 'XWPFNumbering' object when a numbered list is detected in the text, previously it was creating regardless of the text format.
Rearranged the dependencies in 'build.gradle.kts'.
Added 'log4j-core:2.21.0' dependency to improve logging feature and 'junit-platform-launcher' to runtime scope for testing. Updated gradle version from 8.3 to 8.4 for latest features and fixes.
The println statement in 'FileUtilsTest' was updated to print the specific filename 'meeting_minutes.docx'.