Add functionality to list local models
This update introduces the ability to list local models in "OllamaService". The new "listLocalModels" method makes a GET request to the API endpoint and returns a list of models. The models are sorted by modification date and name in a new test in "OllamaServiceTest". An "OffsetDateTimeDeserializer" is also added for JSON deserialization concerns.