Update streaming test and refactor streaming method
The test for the streaming method in OllamaServiceTest has been updated with a new prompt. Moreover, the streaming method in OllamaService has been refactored to simplify the json request creation and the url generation. Now, a new OllamaTextRequest object is created and then it's transformed to JSON using gson.toJson instead of manually constructing it. The url is created by formatting the OLLAMA_BASE_URL constant.