Update libraries versions and refactor code
The libraries' versions for assertj, httpclient, and tika have been updated in the 'libs.versions.toml' file. Additionally, in 'MonitorFineTune.java', the instantiation of 'httpGet' has been refactored to use the 'var' keyword for cleaner code. Also, in 'GeminiApiClientTest.java', the example content string was changed to 'Mockito Made Clear'.